duui-openConverter-reader/requirements.txt
uvicorn
pydantic
fastapi
dkpro-cassis
pydantic_settings
tqdm
requests
python-multipart
websockets